+++ db/migrate/20230118131413_add_spending_breakdown_filename_to_activities.rb
@@ -1,0 +1,5 @@
+class AddSpendingBreakdownFilenameToActivities < ActiveRecord::Migration[6.1]
+  def change
+    add_column :activities, :spending_breakdown_filename, :string
+  end
+end